feat: Add logout and delete account popups
This commit adds two new popups for logging out and deleting the user account. The `openLogoutPopup` and `openDeleteAccountPopup` methods are added to the `PopupManager` class. These methods open the respective popups using the `openCustomPopup` function.

The `DonorSettingsPageController` is updated to use the new popup methods instead of directly calling the `LogoutDialog` and `DeleteAccountDialog` classes.

The `DeleteAccountDialog` class is also updated to use a `SizedBox` and `Padding` to provide better spacing and layout for the dialog content.

This change improves the user experience by providing dedicated popups for logging out and deleting the user account.