feat: Fix backend call in donor personal details save function
The backend call in the save function of the DonorPersonalDetailsController has been fixed to update the user's name and country. The first name and second name text fields are used to construct the name, and the location field is used to determine the country. If an error occurs during the update, it is logged in the controller.