[DPS] DIN-67 Rh donor settings username and password (#26)
* feat: Update donor personal details page

The donor personal details page has been updated to include separate fields for the first name and second name. This change required modifications to the controller and the personal details page. The first name and second name text fields have been added to the personal details page, and the corresponding text controllers have been updated in the controller.

* feat: Separate first name and last name fields in donor personal details page

* feat: Add country field to User model

* feat: Add location field to donor personal details page

* feat: Fix backend call in donor personal details save function

The backend call in the save function of the DonorPersonalDetailsController has been fixed to update the user's name and country. The first name and second name text fields are used to construct the name, and the location field is used to determine the country. If an error occurs during the update, it is logged in the controller.

* feat: Update donor personal details page

The donor personal details page has been updated to include separate fields for the first name and second name. This change required modifications to the controller and the personal details page. The first name and second name text fields have been added to the personal details page, and the corresponding text controllers have been updated in the controller.