DIN-68 [DPS] Donor Settings Logout and Delete Account (#25)
* feat: Add logout and delete account popups

This commit adds two new popups for logging out and deleting the user account. The `openLogoutPopup` and `openDeleteAccountPopup` methods are added to the `PopupManager` class. These methods open the respective popups using the `openCustomPopup` function.

The `DonorSettingsPageController` is updated to use the new popup methods instead of directly calling the `LogoutDialog` and `DeleteAccountDialog` classes.

The `DeleteAccountDialog` class is also updated to use a `SizedBox` and `Padding` to provide better spacing and layout for the dialog content.

This change improves the user experience by providing dedicated popups for logging out and deleting the user account.

* feat: Update donor settings page with new popups

The donor settings page has been updated to use the new `openLogoutPopup` and `openDeleteAccountPopup` methods from the `PopupManager` class. This change replaces the direct usage of the `LogoutDialog` and `DeleteAccountDialog` classes. Additionally, the `DeleteAccountDialog` class has been improved with better spacing and layout using `SizedBox` and `Padding`.

This update enhances the user experience by providing dedicated popups for logging out and deleting the user account.